# Add a benchmark command