--- a/script/ci
+++ b/script/ci
@@ -7,3 +7,4 @@ set -e
 script/fetch-fixtures
 script/check-mallocs
 script/test -b
+script/benchmark