--- a/script/ci
+++ b/script/ci
@@ -2,8 +2,6 @@
 
 set -e
 
-. script/lib.sh
-
 script/fetch-fixtures
 script/check-mallocs
 script/test -b