--- a/script/test
+++ b/script/test
@@ -2,8 +2,6 @@
 
 set -e
 
-. script/lib.sh
-
 function usage {
   cat <<-EOF
 USAGE
@@ -92,6 +90,7 @@ else
 fi
 
 if [[ -n "$run_scan_build" ]]; then
+  . script/util/scan-build.sh
   scan_build make -j2 $target
 else
   make -j2 $target