--- a/script/util/scan-build.sh
+++ b/script/util/scan-build.sh
@@ -1,5 +1,3 @@
-#!/usr/bin/env bash
-
 function scan_build {
   extra_args=()