--- a/test/helpers/random_helpers.h
+++ b/test/helpers/random_helpers.h
@@ -2,7 +2,9 @@
 #define HELPERS_RANDOM_HELPERS_H_
 
 #include <string>
+#include <vector>
 
 std::string random_words(size_t count);
+std::string select_random(const std::vector<std::string> &);
 
 #endif  // HELPERS_RANDOM_HELPERS_H_