--- a/test/helpers/read_test_entries.cc
+++ b/test/helpers/read_test_entries.cc
@@ -92,3 +92,15 @@ vector<TestEntry> read_test_language_corpus(string language_name) {
 
   return result;
 }
+
+vector<ExampleEntry> examples_for_language(string language_name) {
+  vector<ExampleEntry> result;
+  string examples_directory = fixtures_dir + "grammars/" + language_name + "/examples";
+  for (string &filename : list_directory(examples_directory)) {
+    result.push_back({
+      filename,
+      read_file(examples_directory + "/" + filename)
+    });
+  }
+  return result;
+}