--- a/test/helpers/record_alloc.cc
+++ b/test/helpers/record_alloc.cc
@@ -1,7 +1,6 @@
 #include <stdlib.h>
 #include <map>
 #include <set>
-#include "bandit/bandit.h"
 
 using std::map;
 using std::set;