--- a/tests.gyp
+++ b/tests.gyp
@@ -18,11 +18,8 @@
         'test/helpers/load_language.cc',
         'test/helpers/read_test_entries.cc',
         'test/helpers/stderr_logger.cc',
+        'test/helpers/record_alloc.cc',
       ],
-      'configurations': {'Test': {}, 'Release': {}},
-      'default_configuration': 'Release',
-      'xcode_settings': {'CLANG_CXX_LANGUAGE_STANDARD': 'c++11'},
-      'cflags_cc': ['-std=c++11'],
     },
     {
       'target_name': 'tests',
@@ -45,40 +42,34 @@
         '<!@(find test/integration -name "*.cc")',
         '<!@(find test/helpers -name "*.cc")',
       ],
-      'libraries': [
-        '-ldl'
-      ],
-      'default_configuration': 'Test',
-      'configurations': {'Test': {}, 'Release': {}},
-      'cflags': [
-        '-g',
-        '-O0',
+    }
+  ],
+
+  'target_defaults': {
+    'default_configuration': 'Test',
+    'configurations': {'Test': {}, 'Release': {}},
+    'cflags': [
+      '-g',
+      '-O0',
+      '-Wall',
+      '-Wextra',
+      '-Wno-unused-parameter',
+      '-Wno-unknown-pragmas',
+    ],
+    'cflags_cc': ['-std=c++14'],
+    'ldflags': ['-g'],
+    'libraries': ['-ldl'],
+    'xcode_settings': {
+      'CLANG_CXX_LANGUAGE_STANDARD': 'c++14',
+      'OTHER_LDFLAGS': ['-g'],
+      'OTHER_CPLUSPLUSFLAGS': ['-fsanitize=address'],
+      'GCC_OPTIMIZATION_LEVEL': '0',
+      'ALWAYS_SEARCH_USER_PATHS': 'NO',
+      'WARNING_CFLAGS': [
         '-Wall',
         '-Wextra',
-        '-Wno-unused-parameter',
-        '-Wno-unknown-pragmas',
+        '-Wno-unused-parameter'
       ],
-      'cflags_c': [
-        '-std=c99',
-      ],
-      'cflags_cc': [
-        '-std=c++14',
-      ],
-      'ldflags': [
-        '-g',
-      ],
-      'xcode_settings': {
-        'CLANG_CXX_LANGUAGE_STANDARD': 'c++14',
-        'OTHER_LDFLAGS': ['-g'],
-        'OTHER_CPLUSPLUSFLAGS': ['-fsanitize=address'],
-        'GCC_OPTIMIZATION_LEVEL': '0',
-        'ALWAYS_SEARCH_USER_PATHS': 'NO',
-        'WARNING_CFLAGS': [
-          '-Wall',
-          '-Wextra',
-          '-Wno-unused-parameter'
-        ],
-      },
-    }
-  ]
+    },
+  }
 }