--- a/tests.gyp
+++ b/tests.gyp
@@ -2,6 +2,7 @@
   'targets': [
     {
       'target_name': 'benchmarks',
+      'default_configuration': 'Release',
       'type': 'executable',
       'dependencies': [
         'project.gyp:runtime',
@@ -18,11 +19,11 @@
         'test/helpers/load_language.cc',
         'test/helpers/read_test_entries.cc',
         'test/helpers/stderr_logger.cc',
-        'test/helpers/record_alloc.cc',
       ],
     },
     {
       'target_name': 'tests',
+      'default_configuration': 'Test',
       'type': 'executable',
       'dependencies': [
         'project.gyp:runtime',
@@ -42,34 +43,35 @@
         '<!@(find test/integration -name "*.cc")',
         '<!@(find test/helpers -name "*.cc")',
       ],
+      'cflags': [
+        '-g',
+        '-O0',
+        '-Wall',
+        '-Wextra',
+        '-Wno-unused-parameter',
+        '-Wno-unknown-pragmas',
+      ],
+      'ldflags': ['-g'],
+      'xcode_settings': {
+        'OTHER_LDFLAGS': ['-g'],
+        'OTHER_CPLUSPLUSFLAGS': ['-fsanitize=address'],
+        'GCC_OPTIMIZATION_LEVEL': '0',
+        'WARNING_CFLAGS': [
+          '-Wall',
+          '-Wextra',
+          '-Wno-unused-parameter'
+        ],
+      },
     }
   ],
 
   'target_defaults': {
-    'default_configuration': 'Test',
     'configurations': {'Test': {}, 'Release': {}},
-    'cflags': [
-      '-g',
-      '-O0',
-      '-Wall',
-      '-Wextra',
-      '-Wno-unused-parameter',
-      '-Wno-unknown-pragmas',
-    ],
     'cflags_cc': ['-std=c++14'],
-    'ldflags': ['-g'],
     'libraries': ['-ldl'],
     'xcode_settings': {
       'CLANG_CXX_LANGUAGE_STANDARD': 'c++14',
-      'OTHER_LDFLAGS': ['-g'],
-      'OTHER_CPLUSPLUSFLAGS': ['-fsanitize=address'],
-      'GCC_OPTIMIZATION_LEVEL': '0',
       'ALWAYS_SEARCH_USER_PATHS': 'NO',
-      'WARNING_CFLAGS': [
-        '-Wall',
-        '-Wextra',
-        '-Wno-unused-parameter'
-      ],
-    },
+    }
   }
 }