This adds a new command, `script/benchmark` that parses all of the files in the `examples` folders of the fixture grammars. It also parses the examples with the *wrong* grammars, to measure the speed of error recovery in large files with lots of errors.

```sh
c
  cluster.c                     	55 ms		3939 bytes/ms
  malloc.c                      	6 ms		2016 bytes/ms
  parser.c                      	16 ms		2779 bytes/ms
  marker-index.h                	14 ms		345 bytes/ms
  rule.cc                       	18 ms		469 bytes/ms
  jquery.js                     	230 ms		1075 bytes/ms
  python2-grammar.py            	268 ms		115 bytes/ms
  python3-grammar.py            	157 ms		196 bytes/ms

cpp
  marker-index.h                	1 ms		4832 bytes/ms
  rule.cc                       	3 ms		2816 bytes/ms
  cluster.c                     	57 ms		3801 bytes/ms
  malloc.c                      	26 ms		465 bytes/ms
  parser.c                      	21 ms		2118 bytes/ms
  jquery.js                     	228 ms		1084 bytes/ms
  python2-grammar.py            	337 ms		91 bytes/ms
  python3-grammar.py            	306 ms		100 bytes/ms

javascript
  jquery.js                     	55 ms		4497 bytes/ms
  cluster.c                     	324 ms		668 bytes/ms
  malloc.c                      	33 ms		366 bytes/ms
  parser.c                      	68 ms		654 bytes/ms
  marker-index.h                	8 ms		604 bytes/ms
  rule.cc                       	103 ms		82 bytes/ms
  python2-grammar.py            	95 ms		326 bytes/ms
  python3-grammar.py            	102 ms		301 bytes/ms

python
  python2-grammar.py            	10 ms		3098 bytes/ms
  python3-grammar.py            	10 ms		3077 bytes/ms
  cluster.c                     	1232 ms		175 bytes/ms
  malloc.c                      	105 ms		115 bytes/ms
  parser.c                      	137 ms		324 bytes/ms
  marker-index.h                	35 ms		138 bytes/ms
  rule.cc                       	41 ms		206 bytes/ms
  jquery.js                     	868 ms		284 bytes/ms

without errors:
  average speed                 	3381 bytes/ms
  worst speed                   	2016 bytes/ms

with errors:
  average speed                 	587 bytes/ms
  worst speed                   	82 bytes/ms
```